GetOurChainID was being called a lot and logged each call... change it to fetch the configuration value once in init and get rid of some of the debug logging clutter.

+++ src/wallet.cpp
@@ -1045,7 +1045,7 @@ void CWallet::ReserveKeyFromKeyPool(int64& nIndex, CKeyPool& keypool)
         CWalletDB walletdb(strWalletFile);
 
         // Top up key pool
-        int64 nTargetSize = max(GetArg("-keypool", 100), (int64)0);
+        int64 nTargetSize = max(GetArg("-keypool", 100), 0);
         while (setKeyPool.size() < nTargetSize+1)
         {
             int64 nEnd = 1;
